Restructure appearance structs
