+++ Sources/Tabman/TabmanBar/Appearance/Appearance+Interaction.swift
@@ -1,0 +1,18 @@
+//
+//  Appearance+Interaction.swift
+//  Tabman
+//
+//  Created by Merrick Sapsford on 17/05/2017.
+//  Copyright © 2017 Merrick Sapsford. All rights reserved.
+//
+
+import UIKit
+
+// MARK: - TabmanBar interaction appearance properties.
+extension TabmanBar.Appearance {
+    
+    public struct Interaction {
+        /// Whether user scroll is enabled on a scrolling button bar.
+        public var isScrollEnabled: Bool?
+    }
+}
