+++ Sources/Tabman/TabmanBar/Appearance/Appearance+Text.swift
@@ -1,0 +1,18 @@
+//
+//  Appearance+Text.swift
+//  Tabman
+//
+//  Created by Merrick Sapsford on 17/05/2017.
+//  Copyright © 2017 Merrick Sapsford. All rights reserved.
+//
+
+import UIKit
+
+// MARK: - TabmanBar text appearance properties.
+extension TabmanBar.Appearance {
+
+    public struct Text {
+        /// The font to use for text labels in the bar.
+        public var font: UIFont?
+    }
+}
